+++ forui/lib/src/widgets/date_field/calendar/calendar_date_field.dart
@@ -65,6 +65,7 @@ class _CalendarDateField extends FDateField implements FDateFieldCalendarPropert
     super.label,
     super.description,
     super.enabled = true,
+    super.onChange,
     super.onSaved,
     super.autovalidateMode = AutovalidateMode.onUnfocus,
     super.forceErrorText,
@@ -101,6 +102,7 @@ class _CalendarDatePickerState extends _FDateFieldState<_CalendarDateField> {
     super.initState();
     _controller._calendar.addListener(_updateTextController);
     _controller.calendar.addListener(_updateFocus);
+    _controller.addValueListener(widget.onChange);
   }
 
   @override
@@ -120,12 +122,17 @@ class _CalendarDatePickerState extends _FDateFieldState<_CalendarDateField> {
       } else {
         _controller._calendar.removeListener(_updateTextController);
         _controller.calendar.removeListener(_updateFocus);
+        _controller.removeValueListener(old.onChange);
       }
 
       _controller = widget.controller ?? FDateFieldController(vsync: this);
       _controller._calendar.addListener(_updateTextController);
       _controller.calendar.addListener(_updateFocus);
+      _controller.addValueListener(widget.onChange);
       _updateTextController();
+    } else if (widget.onChange != old.onChange) {
+      _controller.removeValueListener(old.onChange);
+      _controller.addValueListener(widget.onChange);
     }
   }
 
@@ -210,6 +217,7 @@ class _CalendarDatePickerState extends _FDateFieldState<_CalendarDateField> {
     } else {
       _controller._calendar.removeListener(_updateTextController);
       _controller.calendar.removeListener(_updateFocus);
+      _controller.removeValueListener(widget.onChange);
     }
 
     if (widget.focusNode == null) {